requirements.txt
langchain==0.3.20
python-dotenv==1.0.1
streamlit==1.43.0
unstructured==0.16.23
tiktoken==0.9.0
faiss-cpu==1.10.0
libmagic==1.0.0
python-magic==0.4.27  
python-magic-bin==0.4.14
groq==0.18.0
nltk==3.9.1
sentence-transformers==3.4.1